+++ Sources/XcodeProj/Objects/Targets/PBXProductType.swift
@@ -28,6 +28,7 @@ public enum PBXProductType: String, Decodable {
     case instrumentsPackage = "com.apple.product-type.instruments-package"
     case intentsServiceExtension = "com.apple.product-type.app-extension.intents-service"
     case onDemandInstallCapableApplication = "com.apple.product-type.application.on-demand-install-capable"
+    case metalLibrary = "com.apple.product-type.metal-library"
 
     /// Returns the file extension for the given product type.
     public var fileExtension: String? {
@@ -56,6 +57,8 @@ public enum PBXProductType: String, Decodable {
             return "instrpkg"
         case .xcFramework:
             return "xcframework"
+        case .metalLibrary:
+            return "metallib"
         case .none:
             return nil
         }